+++ strings/kebab_case.go
@@ -1,0 +1,5 @@
+package strings
+
+func KebabCase(input string) string {
+	return ScreamingDelimited(input, '-', "", false)
+}